Basic Cell-based Eval
Partially addresses #63
Adds two commands: `run-cell` and `run-cell-and-move`. When either command is executed, the code in the current cell is eval'd. Results are displayed block style, below the cell delimiter.
Cells are delimited with `#---` by default, but this also adds an option to set the delimiter in the julia-client settings.

Gotcha: doesn't support anything to do with selections, probably won't do what users expect when the commands added are run when text is selected.